improve UX/DX of auth flow
