+++ docker-compose.yml
@@ -17,6 +17,14 @@ services:
       - POSTGRES_URL=postgresql://postgres:postgres@db:5432/simstudio
       - BETTER_AUTH_URL=http://localhost:3000
       - NEXT_PUBLIC_APP_URL=http://localhost:3000
+      - BETTER_AUTH_SECRET=your_auth_secret_here
+      - ENCRYPTION_KEY=your_encryption_key_here
+      - GOOGLE_CLIENT_ID=placeholder
+      - GOOGLE_CLIENT_SECRET=placeholder
+      - GITHUB_CLIENT_ID=placeholder
+      - GITHUB_CLIENT_SECRET=placeholder
+      - RESEND_API_KEY=placeholder
+      - WEBCONTAINER_CLIENT_ID=placeholder
     depends_on:
       db:
         condition: service_healthy
