+++ sim/app/(auth)/components/oauth-provider-checker.tsx
@@ -1,0 +1,21 @@
+'use server'
+
+export async function getOAuthProviderStatus() {
+  const githubAvailable = !!(
+    process.env.GITHUB_CLIENT_ID &&
+    process.env.GITHUB_CLIENT_SECRET &&
+    process.env.GITHUB_CLIENT_ID !== 'placeholder' &&
+    process.env.GITHUB_CLIENT_SECRET !== 'placeholder'
+  )
+
+  const googleAvailable = !!(
+    process.env.GOOGLE_CLIENT_ID &&
+    process.env.GOOGLE_CLIENT_SECRET &&
+    process.env.GOOGLE_CLIENT_ID !== 'placeholder' &&
+    process.env.GOOGLE_CLIENT_SECRET !== 'placeholder'
+  )
+
+  const isProduction = process.env.NODE_ENV === 'production'
+
+  return { githubAvailable, googleAvailable, isProduction }
+}
