+++ sim/app/lib/auth.ts
@@ -12,9 +12,11 @@ const logger = createLogger('Auth')
 
 // If there is no resend key, it might be a local dev environment
 // In that case, we don't want to send emails and just log them
-const resend =
-  process.env.RESEND_API_KEY && process.env.RESEND_API_KEY.trim() !== ''
-    ? new Resend(process.env.RESEND_API_KEY)
+
+const validResendAPIKEY = process.env.RESEND_API_KEY && process.env.RESEND_API_KEY.trim() !== '' && process.env.RESEND_API_KEY !== 'placeholder'
+
+const resend = validResendAPIKEY
+  ? new Resend(process.env.RESEND_API_KEY)
     : {
         emails: {
           send: async (...args: any[]) => {
@@ -95,7 +97,7 @@ export const auth = betterAuth({
           }
 
           // In development with no RESEND_API_KEY, log verification code
-          if (!process.env.RESEND_API_KEY || process.env.RESEND_API_KEY.trim() === '') {
+          if (!validResendAPIKEY) {
             logger.info('🔑 VERIFICATION CODE FOR LOGIN/SIGNUP', {
               email: data.email,
               otp: data.otp,
